Fix non-devel installs. (#121)
(for #120)

In some cases the wrong files were getting installed.

Also, running from a development env (AKA, git checkout) causes pyperformance to get installed into a venv. However, pip treats this as a "develop" install and the venv basically has a link back to the original files (rather than making a copy). This lead to incorrect logic when detecting "is installed"

Both issues are resolved here.