Support calling `compile` command from when cwd is a source checkout
The `pyperformance compile` command currently fails when the current directory
is a source checkout, since the installation of pyperformance in the benchmark
venv is slightly broken, and the detection of running from a source checkout is
slightly broken. This is true whether using `python -m pyperformance` or `python
./dev.py` methods.

- To detect whether performance is being imported from an "installation" rather
than the source directory, it imports toml and then sees if there is a
`pyperformance` package sitting alongside it. This (perhaps) worked in the past
when `toml` was a module, but it's currently a package, so we need to go up two
levels of directories.

- `pyperformance` is installed in the benchmark venvs with `pip install .`, but
`pip install -e .` actually installs in editable/developer mode, as described
elsewhere.