Allow for specifying local wheels and sdists as dependencies (#215)
It is useful, especially when creating custom versions of dependencies that are compatible with CPython main, to specify local file paths as dependencies. Unfortunately, that's blocked by a feature that treats any dependency that matches a file path as a requirements file. This makes that check a little more specific which allows for installing wheels.