Add a Dask benchmark (gh-246)
The Dask scheduler is largely pure-python-code-bound. There is a discussion about the details here. For this reason, it makes a good real-world workload of a distributed system benchmark (even when the benchmark is running locally on a single core).

Thanks to @TomAugspurger for the suggestion, and @mrocklin for providing the meat of the benchmark itself.