Don't run the daily test on contributors' GitHub forks
I'm currently getting a daily email informing me that the pyperformance tests failed on the `AlexWaygood/pyperformance` repo (because I enabled workflows for my fork). This PR makes it so that the `test` workflow only happens on `python/pyperformance`, not on contributors' forks of the repo (unless they specifically request the workflow by manually triggering the workflow or by making a PR to their fork).