Add a feature for using the same number of loops as a previous run (#327)
Motivation:

On the Faster CPython team, we often collect pystats (counters of various interpreter events) by running the benchmark suite.  It is very useful to compare the stats between two commits to see how a pull request affects the interpreter.  Unfortunately, with pyperformance's default behavior where the number of loops is automatically calibrated, each benchmark may not be run the same number of times from run-to-run, making the data hard to compare.

This change adds a new argument to the "run" command which will use the same number of loops as a previous run.  The loops for each benchmark is looked up from the metadata in the .json output of that previous run, and passed to the underlying call to pyperf using the --loops argument.

Additionally, this modifies one of the benchmarks (sqlglot) to be compatible with that scheme.  sqlglot is the only run_benchmark.py script that runs multiple benchmarks within it in a single call to the script.  This makes it impossible to set the number of loops independently for each of these benchmarks.  It's been updated to use the pattern from other "suites" of benchmarks (e.g. async_tree) where each benchmark has its own .toml file and is run independently.  This should still be backward compatible with older data collected from this benchmark, but doing "pyperformance run -b sqlglot" will now only run a single benchmark.