Update sqlglot to V2
In the current sqlglot normalized, conjunction string would be modified to its normalized form after the first iteration.
And for the remaining iterations, normalization function was applied to the normalized form of the original conjunction string.
Therefore, the performance of the first iteration would be different to the remaining iterations.
This causes unsuitability in performance reports.

To fix this issue, we created the V2 version, in which, a new conjunction string is created for every individual iteration.